refactor: move fields to separate file
